Create a structure Organization with organization name and organization ID as its data members. Next, create another structure Employee that is nested in structure Organization with employee ID, employee salary and employee name as its data members. Write a program in such a way that there are two organizations and each of these contains two employees.
Create a structure Organization with organization name and organization ID as its data
members. Next, create another structure Employee that is nested in structure
Organization with employee ID, employee salary and employee name as its data
members. Write a program in such a way that there are two organizations and each of
these contains two employees.